requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
ccxt==4.4.77
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distro==1.9.0
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
idna==3.10
jiter==0.9.0
multidict==6.4.3
numpy==2.2.5
openai==1.75.0
pandas==2.2.3
pandas_ta==0.3.14b0
propcache==0.3.1
pycares==4.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
